Refactor: Improve date picker dialog and transactions list
- Moved `FiltersDatePickerDialog`, `DatePickerDialogType`, `DatePickerDialogConfig` to the `ui-logic` module.
- Added `FilterDateRangeSelectionData` and `TransactionsFilterableAttributes` data classes.
- Improved handling of errors in `TransactionsInteractor` when getting the transactions.
- Fixed the package of `FiltersSearchBar` component.
- Fixed the package of `FilterDateRangeSelectionData` component.
- Updated `TransactionsScreen` to use `ContentErrorConfig` to display errors.
- Removed unused `DatePickerDialogType` in the `TransactionsViewModel`.
