+++ dashboard-feature/src/main/java/eu/europa/ec/dashboardfeature/model/FilterableTransactionItem.kt
@@ -19,7 +19,7 @@ package eu.europa.ec.dashboardfeature.model
 import eu.europa.ec.businesslogic.validator.model.FilterableAttributes
 import java.time.Instant
 
-internal data class TransactionsFilterableAttributes(
+data class TransactionsFilterableAttributes(
     override val searchTags: List<String>,
     val transactionStatus: TransactionUiStatus,
     val creationDate: Instant?,
